修改删除文章方式

+++ article/views.py
@@ -9,10 +9,13 @@ import markdown
 # Create your views here.
 
 
-def article_delete(request, id):
-    article = ArticlePost.objects.get(id=id)
-    article.delete()
-    return redirect("article:article_list")
+def article_safe_delete(request, id):
+    if request.method == "POST":
+        article = ArticlePost.objects.get(id=id)
+        article.delete()
+        return redirect("article:article_list")
+    else:
+        return HttpResponse("仅允许post请求")
 
 def article_create(request):
     if request.method == "POST":
